doc: Update documentation throughout repository with examples
Primary objective of this commit is to create some top-level documents
in `src/lib.rs` to provide a variety of examples of how to use the
content of this crate.

This commit also includes some documentation for each of the modules to
provide a more clear explanation of the contents of those modules and
how to use them.