bug: Fix issue with infinite recursion when instantiating document
Fixes an issue where the function `models::resource_to_attrs` would
recurse infinitely and crash the tests causing a stack overflow bug.

Fixes the the issue by creating a list of nodes/resource objects that
have been visited when parsing a document and attempting to instantiate
a struct. By tracking the visited nodes and correctly placing some early
return statements we are able to prevent infinite recursion.

This commit also modifies the test that originally identified this
issue. It removes the assertion that a parsed document can be
reserialized into its original form